refactor: move entry point to src/index.js
Put the main client alongside its modules under src/. Updated everything that
referenced it:
- package.json main -> src/index.js
- src/index.js internal requires (./src/x -> ./x)
- bin/wyze-login.js, example/index.js, and the test files (../index.js -> ../src/index.js)
- README run note

No API change: require('wyze-node') still resolves to the Wyze class. Tests
(19/19) and a live signed call both pass after the move.
